+++ Cartfile.private
@@ -1,1 +1,3 @@
 github "typelift/SwiftCheck"
+github "typelift/Operadics"
+
