+++ Cartfile.resolved
@@ -1,1 +1,2 @@
+github "typelift/Operadics" "0.2.3"
 github "typelift/SwiftCheck" "0.8.0"
