+++ Package.pins
@@ -1,0 +1,18 @@
+{
+  "autoPin": true,
+  "pins": [
+    {
+      "package": "Operadics",
+      "reason": null,
+      "repositoryURL": "https://github.com/typelift/Operadics.git",
+      "version": "0.2.3"
+    },
+    {
+      "package": "SwiftCheck",
+      "reason": null,
+      "repositoryURL": "https://github.com/typelift/SwiftCheck.git",
+      "version": "0.8.0"
+    }
+  ],
+  "version": 1
+}
