+++ Package.swift
@@ -1,10 +1,20 @@
 // swift-tools-version:3.1
 
+import Foundation
 import PackageDescription
+
+// HACK from https://github.com/ReactiveCocoa/ReactiveSwift/blob/master/Package.swift
+var isSwiftPMTest: Bool {
+    return ProcessInfo.processInfo.environment["SWIFTPM_TEST_Abstract"] == "YES"
+}
 
 let package = Package(
     name: "Abstract",
-    exclude: [
-		"Tests"
-	]
+    dependencies: [
+        .Package(url: "https://github.com/typelift/Operadics.git",
+                 majorVersion: 0)
+    ] + (isSwiftPMTest ?
+      [.Package(url: "https://github.com/typelift/SwiftCheck.git", versions: Version(0,6,0)..<Version(1,0,0))] :
+      [])
 )
+
